[androd] Moved DialogUtils to dialog package
